+++ client/CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- Fixed the timeline list accessibility ([#865](https://github.com/project-lux/lux-frontend/issues/865)).
+
 ### Removed
 
 ## v1.74.1 - 2026-07-20